requirements.txt
chainlit==1.0.400
langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.1
tiktoken==0.5.2
fastapi
uvicorn
python-multipart
groq
pytesseract
pdf2image
Pillow
rank-bm25==0.2.2
pymupdf==1.26.6
langchain-huggingface
